feat: Implement first prototype for translating parsed query into db ops
This is the first prototype that is able to translate the parsed query into
a Prisma ORM filtering structure to be executed.

But the implementation will need to change to use a raw SQL query instead,
because my search syntax allows for more complex stuff like RegEx.
It looked like performance using Prisma ORM was a very little concern – I expected it to be a lot worse