feat: Implement support for incremental project indexing
Updating the Repository relation itself within the database transaction has been
moved to the end of it.
This is done to prevent a row-level lock on the repository slowing down/preventing
the login and/or updating of the user's permissions.

To the transaction itself, it doesn't make any difference - We just have to make
sure the project already exists in the database.