feat: Chunk file content in database and configure GIST index
The GIST index speeds up search queries by **a lot**.
But one of its limitations is the maximum number of "input bytes".
That's why we chunk our file content to not exceed that limit of 8191 bytes.

To fully support Unicode strings we split the file content string based on
unicode code points instead of using `Buffer#subarray` which would produce
the chunks but *might* cause multi-byte chars to be split and not being searchable.